test(io): move logs and observations tests to test_log.py
